requirements.txt
Django==3.0.8
psycopg2==2.8.5
dj-database-url==0.5.0
python-dotenv==0.14.0